samples/pipeline/python/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
../../../packages/pdk-nag/dist/python/aws_prototyping_sdk.pdk_nag-0.0.0-py3-none-any.whl
../../../packages/pipeline/dist/python/aws_prototyping_sdk.pipeline-0.0.0-py3-none-any.whl
aws-cdk-lib
constructs
pyhumps
